# Speed Up Packed Compression