`np.(un)pack_bits` and the subsequent array manipulation for 4bit packed compression was really slow for larger models. I re-implemented both packing and unpacking using bit manipulation instead and its a lot faster. This also greatly simplified the 4bit vs 8bit packing code

#### Before
* packing for 1.1b: 72 sec
* unpacking for 1.1: 40 sec

#### After
* packing for 1.1b: 3 sec
* unpacking for 1.1: 7 sec

### Testing
Confirmed the packing output is indentical to the previous implementation, and added an additional unit test for int8 packing